+++ src/CheckIn.js
@@ -1,14 +1,19 @@
-import { get } from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Field, reduxForm } from 'redux-form';
+import { Field } from 'react-final-form';
 import {
   FormattedMessage,
   FormattedTime,
   injectIntl,
-  intlShape
 } from 'react-intl';
 import moment from 'moment-timezone';
+import stripesFinalForm from '@folio/stripes/final-form';
+import createInactivityTimer from 'inactivity-timer';
+import {
+  get,
+  isEmpty,
+} from 'lodash';
+
 import {
   Paneset,
   Pane,
@@ -34,13 +39,16 @@ import { IfPermission } from '@folio/stripes/core';
 import PrintButton from './components/PrintButton';
 import FeesFinesOwnedStatus from './components/FeesFinesOwnedStatus';
 import FeeFineDetailsButton from './components/FeeFineDetailsButton';
-import { convertToSlipData } from './util';
+import {
+  convertToSlipData,
+  getCheckinSettings,
+} from './util';
 import styles from './checkin.css';
 
 class CheckIn extends React.Component {
   static propTypes = {
     stripes: PropTypes.object.isRequired,
-    intl: intlShape.isRequired,
+    intl: PropTypes.object,
     scannedItems: PropTypes.arrayOf(PropTypes.object),
     showCheckinNotes: PropTypes.func,
     handleSubmit: PropTypes.func.isRequired,
@@ -49,9 +57,19 @@ class CheckIn extends React.Component {
     barcodeRef: PropTypes.shape({
       current: PropTypes.instanceOf(Element)
     }),
+    formRef: PropTypes.shape({
+      current: PropTypes.instanceOf(Element)
+    }),
     onSessionEnd: PropTypes.func,
-    change: PropTypes.func,
-    resources: PropTypes.object,
+    resources: PropTypes.shape({
+      checkinSettings: PropTypes.shape({
+        records: PropTypes.arrayOf(PropTypes.object),
+      }),
+      scannedItems: PropTypes.arrayOf(PropTypes.object),
+      staffSlips: PropTypes.shape({
+        records: PropTypes.arrayOf(PropTypes.object),
+      }),
+    }),
     mutator: PropTypes.shape({
       query: PropTypes.shape({
         update: PropTypes.func,
@@ -61,14 +79,15 @@ class CheckIn extends React.Component {
       }),
     }).isRequired,
     loading: PropTypes.bool.isRequired,
+    form: PropTypes.object.isRequired,
   };
 
   constructor(props) {
     super(props);
-    this.onSubmit = this.onSubmit.bind(this);
     this.showInfo = this.showInfo.bind(this);
     this.renderActions = this.renderActions.bind(this);
     this.handleOptionsChange = this.handleOptionsChange.bind(this);
+    this.timer = undefined;
 
     this.state = {
       showPickers: false
@@ -76,15 +95,53 @@ class CheckIn extends React.Component {
   }
 
   componentDidMount() {
-    this.focusInput();
+    this.props.formRef.current = this.props.form;
+  }
+
+  componentDidUpdate() {
+    const {
+      resources: {
+        checkinSettings,
+        checkinSettings: {
+          records: checkinSettingsRecords,
+        },
+        scannedItems,
+      },
+      onSessionEnd,
+    } = this.props;
+
+    if (this.timer) {
+      return;
+    }
+
+    if (!checkinSettings || isEmpty(checkinSettingsRecords)) {
+      return;
+    }
+
+    const parsed = getCheckinSettings(checkinSettingsRecords);
+
+    if (!parsed.checkoutTimeout) {
+      this.timer = null;
+      return;
+    }
+
+    if (scannedItems.length) {
+      this.timer = createInactivityTimer(`${parsed.checkoutTimeoutDuration}m`, () => {
+        onSessionEnd();
+      });
+
+      ['keydown', 'mousedown'].forEach((event) => {
+        document.addEventListener(event, () => {
+          if (this.timer) {
+            this.timer.signal();
+          }
+        });
+      });
+    }
   }
 
   focusInput() {
     this.props.barcodeRef.current.focus();
-  }
-
-  onSubmit(data) {
-    return this.props.submithandler(data, this);
   }
 
   handleSessionEnd = async () => {
@@ -95,7 +152,7 @@ class CheckIn extends React.Component {
   }
 
   showPickers = () => {
-    const { change, intl: { timeZone } } = this.props;
+    const { form: { change }, intl: { timeZone } } = this.props;
     const now = moment.tz(timeZone);
     change('item.checkinDate', now.format());
     change('item.checkinTime', now.format());
@@ -336,11 +393,17 @@ class CheckIn extends React.Component {
     const {
       handleSubmit,
       intl: { formatDate, formatMessage, formatTime },
+      form,
       scannedItems,
       pristine,
       barcodeRef,
       loading,
     } = this.props;
+
+    const {
+      hasSubmitErrors = false,
+      submitErrors = {},
+    } = form.getState();
 
     const { showPickers } = this.state;
     const itemListFormatter = {
@@ -409,7 +472,7 @@ class CheckIn extends React.Component {
     const emptyMessage = !loading ? <FormattedMessage id="ui-checkin.noItems" /> : null;
 
     return (
-      <form onSubmit={handleSubmit(this.onSubmit)}>
+      <form onSubmit={handleSubmit}>
         <div style={containerStyle}>
           <Paneset static>
             <Pane paneTitle={scannedItemsLabel} defaultWidth="100%">
@@ -428,6 +491,7 @@ class CheckIn extends React.Component {
                         component={TextField}
                         data-test-check-in-barcode
                       />
+                      {hasSubmitErrors && <span className={styles.error}>{submitErrors.checkin}</span>}
                     </Layout>
                   </Col>
                   <Col xs={3} sm={1}>
@@ -529,6 +593,6 @@ class CheckIn extends React.Component {
   }
 }
 
-export default reduxForm({
-  form: 'CheckIn',
+export default stripesFinalForm({
+  navigationCheck: true,
 })(injectIntl(CheckIn));