+++ src/ModalManager.js
@@ -8,7 +8,6 @@ import PropTypes from 'prop-types';
 import SafeHTMLMessage from '@folio/react-intl-safe-html';
 import {
   FormattedMessage,
-  intlShape,
   injectIntl,
   FormattedDate,
   FormattedTime,
@@ -24,7 +23,7 @@ import css from './ModalManager.css';
 
 class ModalManager extends React.Component {
   static propTypes = {
-    intl: intlShape,
+    intl: PropTypes.object,
     checkedinItem: PropTypes.object.isRequired,
     checkinNotesMode: PropTypes.bool,
     onDone: PropTypes.func.isRequired,
@@ -37,6 +36,10 @@ class ModalManager extends React.Component {
     const { checkedinItem, checkinNotesMode } = props;
     this.state = { checkedinItem, checkinNotesMode };
     this.steps = [
+      {
+        validate: this.shouldWithdrawnModalBeShown,
+        exec: () => this.setState({ showWithdrawnModal: true }),
+      },
       {
         validate: this.shouldDeclaredLostModalBeShown,
         exec: () => this.setState({ showDeclaredLostModal: true }),
@@ -75,6 +78,11 @@ class ModalManager extends React.Component {
     return this.props.onDone();
   }
 
+  shouldWithdrawnModalBeShown = () => {
+    const { checkedinItem } = this.state;
+    return checkedinItem?.status?.name === statuses.WITHDRAWN;
+  }
+
   shouldDeclaredLostModalBeShown = () => {
     const { checkedinItem } = this.state;
     return get(checkedinItem, 'status.name') === statuses.DECLARED_LOST;
@@ -108,16 +116,20 @@ class ModalManager extends React.Component {
     );
   }
 
+  confirmWithdrawnModal = () => {
+    this.setState({ showWithdrawnModal: false }, () => this.execSteps(1));
+  }
+
   confirmDeclareLostModal = () => {
-    this.setState({ showDeclaredLostModal: false }, () => this.execSteps(1));
+    this.setState({ showDeclaredLostModal: false }, () => this.execSteps(2));
   }
 
   confirmMultipieceModal = () => {
-    this.setState({ showMultipieceModal: false }, () => this.execSteps(2));
+    this.setState({ showMultipieceModal: false }, () => this.execSteps(3));
   }
 
   confirmMissingModal = () => {
-    this.setState({ showMissingModal: false }, () => this.execSteps(3));
+    this.setState({ showMissingModal: false }, () => this.execSteps(4));
   }
 
   confirmCheckinNoteModal = () => {
@@ -131,40 +143,72 @@ class ModalManager extends React.Component {
       showMissingModal: false,
       showMultipieceModal: false,
       showDeclaredLostModal: false,
+      showWithdrawnModal: false,
     });
 
     this.props.onCancel();
   }
 
-  renderDeclaredLostModal() {
-    const {
-      checkedinItem,
-      showDeclaredLostModal,
-    } = this.state;
+  getItemValues = (item) => {
     const {
       barcode,
       title,
       materialType,
-    } = checkedinItem;
+    } = item;
+
+    return {
+      title,
+      barcode,
+      materialType: upperFirst(materialType?.name ?? ''),
+    };
+  }
+
+  renderConfirmStatusModal = (isOpen, status, messageId, onConfirm) => {
+    const { checkedinItem } = this.state;
 
     return (
       <ConfirmationModal
-        id="test-declared-lost-modal"
-        open={showDeclaredLostModal}
+        id={`test-${status}-modal`}
+        open={isOpen}
         item={checkedinItem}
-        heading={<FormattedMessage id="ui-checkin.declaredLostModal.heading" />}
+        heading={<FormattedMessage id={`ui-checkin.${status}Modal.heading`} />}
         message={<SafeHTMLMessage
-          id="ui-checkin.declaredLostModal.message"
-          values={{
-            title,
-            barcode,
-            materialType: upperFirst(materialType?.name ?? ''),
-          }}
+          id={messageId}
+          values={this.getItemValues(checkedinItem)}
         />}
-        onConfirm={this.confirmDeclareLostModal}
+        onConfirm={onConfirm}
         onCancel={this.onCancel}
         confirmLabel={<FormattedMessage id="ui-checkin.statusModal.confirm" />}
       />
+    );
+  }
+
+  renderWithdrawnModal() {
+    const {
+      showWithdrawnModal,
+      checkedinItem,
+    } = this.state;
+    const messageId = checkedinItem.discoverySuppress ?
+      'ui-checkin.withdrawnModal.suppressedMessage' :
+      'ui-checkin.withdrawnModal.notSuppressedMessage';
+
+    return this.renderConfirmStatusModal(
+      showWithdrawnModal,
+      'withdrawn',
+      messageId,
+      this.confirmWithdrawnModal,
+    );
+  }
+
+  renderDeclaredLostModal() {
+    const { showDeclaredLostModal } = this.state;
+    const messageId = 'ui-checkin.declaredLostModal.message';
+
+    return this.renderConfirmStatusModal(
+      showDeclaredLostModal,
+      'declaredLost',
+      messageId,
+      this.confirmDeclareLostModal,
     );
   }
 
@@ -314,6 +358,7 @@ class ModalManager extends React.Component {
       showCheckinNoteModal,
       showMultipieceModal,
       showDeclaredLostModal,
+      showWithdrawnModal
     } = this.state;
 
     return (
@@ -322,6 +367,7 @@ class ModalManager extends React.Component {
         {showCheckinNoteModal && this.renderCheckinNoteModal()}
         {showMultipieceModal && this.renderMultipieceModal()}
         {showDeclaredLostModal && this.renderDeclaredLostModal()}
+        {showWithdrawnModal && this.renderWithdrawnModal()}
       </>
     );
   }