+++ test/bigtest/interactors/check-in.js
@@ -19,6 +19,7 @@ import CheckinNoteModalInteractor from './checkin-note-modal';
 import MissingItemModalInteractor from './missing-item-modal';
 import MultiPieceModalInteractor from './multi-piece-modal';
 import DeclaredLostModalInteractor from './declared-lost-modal';
+import WithdrawnModalInteractor from './withdrawn-modal';
 
 @interactor class CheckInInteractor {
   processDate = new DatepickerInteractor('[data-test-process-date]');
@@ -28,6 +29,8 @@ import DeclaredLostModalInteractor from './declared-lost-modal';
   missingItemModal = new MissingItemModalInteractor();
   checkinNoteModal = new CheckinNoteModalInteractor();
   declaredLostModal = new DeclaredLostModalInteractor();
+  withdrawnModal = new WithdrawnModalInteractor();
+
   selectEllipse = clickable('[data-test-elipse-select] button');
   checkedInItemsList = scoped('#list-items-checked-in', MultiColumnListInteractor);
   selectLoanDetails = clickable('[data-test-loan-details] a');
@@ -35,7 +38,7 @@ import DeclaredLostModalInteractor from './declared-lost-modal';
   selectItemDetails = clickable('[data-test-item-details] a');
   selectRequestDetails = clickable('[data-test-request-details] a');
   selectFeeFineDetails = clickable('[data-test-fee-fine-details] a');
-  fillOutError = text('[data-test-check-in-scan] [class^="feedbackError"]');
+  fillOutError = text('[data-test-check-in-scan] [class^="error"]');
   barcodePresent = isPresent('[data-test-check-in-barcode]');
   barcodeInputValue = value('[data-test-check-in-barcode]');
   barcodeInputIsFocused = isPresent('[data-test-check-in-barcode]:focus');